+++ lib/toonx.dart
@@ -5,3 +5,4 @@ export 'src/decode.dart';
 export 'src/options.dart';
 export 'src/types.dart';
 export 'src/utils.dart';
+export 'src/yaml_converter.dart';